WIP converting privdata to bytestring
The encoding conversion around the read/show in setPrivDataTo and
getLocalPrivData might not be quite right. It's certainly a really nasty
round-tripping to String and back to handle serialization.

Everything that uses withPrivData would need to be updated to use a
ByteString. This is problimatic since a lot of it wants to operate on it by
lines, pass it into other properties that take Strings, etc.