more work on type level port conflict detection
Main problem now is difficulty parameterizing properties that take a port
parameter, and so should have different types depending on the input port.

I think that the way to do that would be to specify the port in the type
of the property. For example:

apache :: Property (UsingPort 80 + DebianLike)
apache = propertyWithPorts "apache" $ \ports -> ...

So, propertyWithMetaPorts would take the metatypes singleton of the
property, and use fromSing to convert it to a value, and filter it down
to just the list of ports.

But, MetaTypes currently has:

        -- undefined because fromSing yields an Integer here, and there's
        -- no way to get from there to a Nat.
        fromSing (UsingPortS _) = UsingPort' undefined